updated stack trace generator to use generic field type

+++ src/main.rs
@@ -94,9 +94,9 @@ fn generate_fibonacci_program(n: usize) -> Vec<u64> {
     // the last operation pops top 2 stack items, adds them, and pushes
     // the result back onto the stack
     for _ in 0..(n - 1) {
-        program.push(opcodes::DUP0);
-        program.push(opcodes::PULL2);
-        program.push(opcodes::ADD);
+        program.push(opcodes::DUP0 as u64);
+        program.push(opcodes::PULL2 as u64);
+        program.push(opcodes::ADD as u64);
     }
 
     return processor::pad_program(&program);
